migrate setup project
