+++ src/Elf.Setup/SetupHelper.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Data;
-using System.IO;
-using System.Reflection;
 using Dapper;
 
 namespace Elf.Setup
@@ -21,38 +19,6 @@ namespace Elf.Setup
                                 "FROM INFORMATION_SCHEMA.TABLES " +
                                 "WHERE TABLE_NAME = N'Link'") == 1;
             return !tableExists;
-        }
-
-        public bool TestDatabaseConnection(Action<Exception> errorLogAction = null)
-        {
-            try
-            {
-                int result = conn.ExecuteScalar<int>("SELECT 1");
-                return result == 1;
-            }
-            catch (Exception e)
-            {
-                errorLogAction?.Invoke(e);
-                return false;
-            }
-        }
-
-        public void SetupDatabase()
-        {
-            var sql = GetEmbeddedSqlScript("schema-mssql-140");
-            if (!string.IsNullOrWhiteSpace(sql))
-            {
-                conn.Execute(sql);
-            }
-        }
-
-        private static string GetEmbeddedSqlScript(string scriptName)
-        {
-            var assembly = typeof(SetupHelper).GetTypeInfo().Assembly;
-            using var stream = assembly.GetManifestResourceStream($"Elf.Setup.Data.{scriptName}.sql");
-            using var reader = new StreamReader(stream);
-            var sql = reader.ReadToEnd();
-            return sql;
         }
     }
 }
