+++ src/Elf.sln
@@ -18,11 +18,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".Solution Items", ".Solutio
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Elf.Services", "Elf.Services\Elf.Services.csproj", "{31EC911F-E050-41D5-BB22-3D193C99FF50}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{26C1CE24-58B3-4D30-9FC0-B9EC35BDE86B}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Application", "Application", "{3FD57E01-5791-42AC-BA79-384D6CC46501}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Elf.Setup", "Elf.Setup\Elf.Setup.csproj", "{3BA4D966-B3E6-43FE-9F71-35A92A0570FE}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{A3995125-2A8A-4BD8-BB11-BECE6DBCA76A}"
 EndProject
@@ -42,10 +38,6 @@ Global
 		{31EC911F-E050-41D5-BB22-3D193C99FF50}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{31EC911F-E050-41D5-BB22-3D193C99FF50}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{31EC911F-E050-41D5-BB22-3D193C99FF50}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3BA4D966-B3E6-43FE-9F71-35A92A0570FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3BA4D966-B3E6-43FE-9F71-35A92A0570FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3BA4D966-B3E6-43FE-9F71-35A92A0570FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3BA4D966-B3E6-43FE-9F71-35A92A0570FE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5EA676F8-A271-40E2-BCBF-09E80E5925C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5EA676F8-A271-40E2-BCBF-09E80E5925C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5EA676F8-A271-40E2-BCBF-09E80E5925C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -56,8 +48,7 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A5F03891-9A66-4CF1-9BA9-D23C9B22AE6E} = {3FD57E01-5791-42AC-BA79-384D6CC46501}
-		{31EC911F-E050-41D5-BB22-3D193C99FF50} = {26C1CE24-58B3-4D30-9FC0-B9EC35BDE86B}
-		{3BA4D966-B3E6-43FE-9F71-35A92A0570FE} = {26C1CE24-58B3-4D30-9FC0-B9EC35BDE86B}
+		{31EC911F-E050-41D5-BB22-3D193C99FF50} = {3FD57E01-5791-42AC-BA79-384D6CC46501}
 		{5EA676F8-A271-40E2-BCBF-09E80E5925C6} = {A3995125-2A8A-4BD8-BB11-BECE6DBCA76A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
